+++ test/EBridge.Contracts.Bridge.Tests/BridgeContractTests.cs
@@ -19,18 +19,8 @@ using DeleteAdminsInput = EBridge.Contracts.Oracle.DeleteAdminsInput;
 
 namespace EBridge.Contracts.Bridge;
 
-public partial class BridgeContractTests : BridgeContractTestBase
+public class BridgeContractTests : BridgeContractTestBase
 {
-    private readonly Address _regimentAddress =
-        Address.FromBase58("2aT9rHLuFRFCHJ1cBSTDR8oD1EFFEBqqiw8fdXD8UuEKRj6Tfh");
-
-    private Hash _swapHashOfElf;
-    private Hash _swapHashOfUsdt;
-    private Hash _swapOfElfSpaceId;
-    private Hash _swapOfUsdtSpaceId;
-
-    private Dictionary<string, Hash> _receiptDictionary;
-
     #region Permission
 
     [Fact]
@@ -283,366 +273,6 @@ public partial class BridgeContractTests : BridgeContractTestBase
     #endregion
 
     #region Helper
-
-    private async Task CreateRegimentTest()
-    {
-        // Create regiment.
-        var executionResult = await OracleContractStub.CreateRegiment.SendAsync(new CreateRegimentInput
-        {
-            Manager = DefaultSenderAddress,
-            InitialMemberList = {Transmitters.Select(a => a.Address)}
-        });
-
-        var regimentAddress = RegimentCreated.Parser
-            .ParseFrom(executionResult.TransactionResult.Logs.First(l => l.Name == nameof(RegimentCreated))
-                .NonIndexed).RegimentAddress;
-        regimentAddress.ShouldBe(_regimentAddress);
-        var regimentInfo = await RegimentContractStub.GetRegimentInfo.CallAsync(_regimentAddress);
-        var manager = regimentInfo.Manager;
-        manager.ShouldBe(DefaultSenderAddress);
-
-        await OracleContractStub.AddAdmins.SendAsync(new AddAdminsInput
-        {
-            RegimentAddress = _regimentAddress,
-            OriginSenderAddress = manager,
-            NewAdmins = {BridgeContractAddress}
-        });
-    }
-
-    private async Task<(Address, Address)> CreateOrganizationTest()
-    {
-        var executionResult = await AssociationContractStub.CreateOrganization.SendAsync(new CreateOrganizationInput
-        {
-            OrganizationMemberList = new OrganizationMemberList
-            {
-                OrganizationMembers =
-                {
-                    RestartNodes[0].Address,
-                    RestartNodes[1].Address,
-                    RestartNodes[2].Address,
-                    RestartNodes[3].Address,
-                    RestartNodes[4].Address,
-                }
-            },
-            ProposalReleaseThreshold = new ProposalReleaseThreshold
-            {
-                MinimalApprovalThreshold = 4,
-                MaximalRejectionThreshold = 1,
-                MinimalVoteThreshold = 4
-            },
-            CreationToken = HashHelper.ComputeFrom("restart"),
-            ProposerWhiteList = new ProposerWhiteList
-            {
-                Proposers =
-                {
-                    RestartNodes[0].Address,
-                    RestartNodes[1].Address,
-                    RestartNodes[2].Address,
-                    RestartNodes[3].Address,
-                    RestartNodes[4].Address,
-                }
-            }
-        });
-        var organizationAddress = (OrganizationCreated.Parser.ParseFrom(executionResult.TransactionResult.Logs
-            .FirstOrDefault(e => e.Name == nameof(OrganizationCreated))?.NonIndexed)).OrganizationAddress;
-        var executionResult1 = await AssociationContractStub.CreateOrganization.SendAsync(new CreateOrganizationInput
-        {
-            OrganizationMemberList = new OrganizationMemberList
-            {
-                OrganizationMembers =
-                {
-                    DefaultAccount.Address,
-                    organizationAddress
-                }
-            },
-            ProposalReleaseThreshold = new ProposalReleaseThreshold
-            {
-                MinimalApprovalThreshold = 2,
-                MaximalRejectionThreshold = 0,
-                MinimalVoteThreshold = 2
-            },
-            CreationToken = HashHelper.ComputeFrom("restart"),
-            ProposerWhiteList = new ProposerWhiteList
-            {
-                Proposers =
-                {
-                    DefaultAccount.Address
-                }
-            }
-        });
-        var organizationAddress1 = (OrganizationCreated.Parser.ParseFrom(executionResult1.TransactionResult.Logs
-            .FirstOrDefault(e => e.Name == nameof(OrganizationCreated))?.NonIndexed)).OrganizationAddress;
-        return (organizationAddress, organizationAddress1);
-    }
-
-    private async Task<(Address, Address)> CreateOrganizationSecondTest()
-    {
-        var executionResult = await AssociationContractStub.CreateOrganization.SendAsync(new CreateOrganizationInput
-        {
-            OrganizationMemberList = new OrganizationMemberList
-            {
-                OrganizationMembers =
-                {
-                    RestartNodes[0].Address,
-                    RestartNodes[1].Address,
-                    RestartNodes[2].Address,
-                    RestartNodes[3].Address,
-                    RestartNodes[4].Address,
-                }
-            },
-            ProposalReleaseThreshold = new ProposalReleaseThreshold
-            {
-                MinimalApprovalThreshold = 4,
-                MaximalRejectionThreshold = 1,
-                MinimalVoteThreshold = 4
-            },
-            CreationToken = HashHelper.ComputeFrom("test"),
-            ProposerWhiteList = new ProposerWhiteList
-            {
-                Proposers =
-                {
-                    RestartNodes[0].Address,
-                    RestartNodes[1].Address,
-                    RestartNodes[2].Address,
-                    RestartNodes[3].Address,
-                    RestartNodes[4].Address,
-                }
-            }
-        });
-        var organizationAddress = (OrganizationCreated.Parser.ParseFrom(executionResult.TransactionResult.Logs
-            .FirstOrDefault(e => e.Name == nameof(OrganizationCreated))?.NonIndexed)).OrganizationAddress;
-        var executionResult1 = await AssociationContractStub.CreateOrganization.SendAsync(new CreateOrganizationInput
-        {
-            OrganizationMemberList = new OrganizationMemberList
-            {
-                OrganizationMembers =
-                {
-                    DefaultAccount.Address,
-                    organizationAddress
-                }
-            },
-            ProposalReleaseThreshold = new ProposalReleaseThreshold
-            {
-                MinimalApprovalThreshold = 2,
-                MaximalRejectionThreshold = 0,
-                MinimalVoteThreshold = 2
-            },
-            CreationToken = HashHelper.ComputeFrom("ttt"),
-            ProposerWhiteList = new ProposerWhiteList
-            {
-                Proposers =
-                {
-                    DefaultAccount.Address
-                }
-            }
-        });
-        var organizationAddress1 = (OrganizationCreated.Parser.ParseFrom(executionResult1.TransactionResult.Logs
-            .FirstOrDefault(e => e.Name == nameof(OrganizationCreated))?.NonIndexed)).OrganizationAddress;
-        return (organizationAddress, organizationAddress1);
-    }
-
-    private async Task InitialOracleContractAsync()
-    {
-        await OracleContractStub.Initialize.SendAsync(new Oracle.InitializeInput
-        {
-            RegimentContractAddress = RegimentContractAddress
-        });
-    }
-
-    private async Task<(Address, Address)> InitialBridgeContractAsync()
-    {
-        var organizationAddress = await CreateOrganizationTest();
-        await BridgeContractStub.Initialize.SendAsync(new InitializeInput
-        {
-            MerkleTreeContractAddress = MerkleTreeContractAddress,
-            OracleContractAddress = OracleContractAddress,
-            RegimentContractAddress = RegimentContractAddress,
-            ReportContractAddress = ReportContractAddress,
-            Admin = DefaultSenderAddress,
-            Controller = DefaultSenderAddress,
-            OrganizationAddress = organizationAddress.Item2,
-            PauseController = DefaultSenderAddress,
-            ApproveTransferController = DefaultSenderAddress
-        });
-        return organizationAddress;
-    }
-
-    private async Task InitialMerkleTreeContractAsync()
-    {
-        await MerkleTreeContractStub.Initialize.SendAsync(new MerkleTreeContract.InitializeInput
-        {
-            Owner = BridgeContractAddress,
-            RegimentContractAddress = RegimentContractAddress
-        });
-    }
-
-    private async Task InitialReportContractAsync()
-    {
-        await PortTokenCreate();
-        await ReportContractStub.Initialize.SendAsync(new Report.InitializeInput
-        {
-            OracleContractAddress = OracleContractAddress,
-            RegimentContractAddress = RegimentContractAddress,
-            ReportFee = 0,
-            InitialRegisterWhiteList = {DefaultSenderAddress}
-        });
-    }
-
-    private async Task CreateAndIssueUSDTAsync()
-    {
-        await TokenContractStub.Create.SendAsync(new CreateInput
-        {
-            Symbol = "USDT",
-            Decimals = 8,
-            Issuer = DefaultSenderAddress,
-            TokenName = "Stable coin",
-            TotalSupply = 10_00000000_00000000,
-            IsBurnable = true
-        });
-
-        await TokenContractStub.Issue.SendAsync(new IssueInput
-        {
-            Symbol = "USDT",
-            Amount = 10_00000000_00000000,
-            To = DefaultSenderAddress
-        });
-    }
-
-    private async Task PortTokenCreate()
-    {
-        // Create PORT token.
-        await TokenContractStub.Create.SendAsync(new CreateInput
-        {
-            TokenName = "Port Token",
-            Decimals = 8,
-            Issuer = DefaultSenderAddress,
-            IsBurnable = true,
-            Symbol = "PORT",
-            TotalSupply = 10_00000000_00000000
-        });
-
-        // Issue PORT token.
-        await TokenContractStub.Issue.SendAsync(new IssueInput
-        {
-            To = Transmitters.First().Address,
-            Amount = 10_00000000_00000000,
-            Symbol = "PORT"
-        });
-
-        // Approve Oracle Contract.
-        var transmitterTokenContractStub = GetTokenContractStub(Transmitters.First().KeyPair);
-        await transmitterTokenContractStub.Approve.SendAsync(new ApproveInput
-        {
-            Symbol = "PORT",
-            Amount = 5_00000000_00000000,
-            Spender = OracleContractAddress
-        });
-    }
-
-    private async Task InitialSetGas()
-    {
-        await BridgeContractStub.SetGasLimit.SendAsync(new SetGasLimitInput
-        {
-            GasLimitList =
-            {
-                new GasLimit
-                {
-                    ChainId = "Ethereum",
-                    GasLimit_ = 293414
-                }
-            }
-        });
-        await BridgeContractStub.SetGasPrice.SendAsync(new SetGasPriceInput
-        {
-            GasPriceList =
-            {
-                new GasPrice
-                {
-                    ChainId = "Ethereum",
-                    GasPrice_ = 8245816000
-                }
-            }
-        });
-        //var priceRatioReal = 1 / 0.000095;
-        await BridgeContractStub.SetPriceRatio.SendAsync(new SetRatioInput
-        {
-            Value =
-            {
-                new Ratio
-                {
-                    ChainId = "Ethereum",
-                    Ratio_ = 1052631578947
-                }
-            }
-        });
-        await BridgeContractStub.SetFeeFloatingRatio.SendAsync(new SetRatioInput
-        {
-            Value =
-            {
-                new Ratio
-                {
-                    ChainId = "Ethereum",
-                    Ratio_ = 20
-                }
-            }
-        });
-        await BridgeContractStub.SetGasLimit.SendAsync(new SetGasLimitInput
-        {
-            GasLimitList =
-            {
-                new GasLimit
-                {
-                    ChainId = "Ploygon",
-                    GasLimit_ = 293414
-                }
-            }
-        });
-        await BridgeContractStub.SetGasPrice.SendAsync(new SetGasPriceInput
-        {
-            GasPriceList =
-            {
-                new GasPrice
-                {
-                    ChainId = "Ploygon",
-                    GasPrice_ = 8245816000
-                }
-            }
-        });
-        await BridgeContractStub.SetPriceRatio.SendAsync(new SetRatioInput
-        {
-            Value =
-            {
-                new Ratio
-                {
-                    ChainId = "Ploygon",
-                    Ratio_ = 1052631578947
-                }
-            }
-        });
-    }
-
-    private async Task CheckBalanceAsync(Address ownerAddress, string symbol, long supposedBalance)
-    {
-        var balance = (await TokenContractStub.GetBalance.CallAsync(new GetBalanceInput
-        {
-            Owner = ownerAddress,
-            Symbol = symbol
-        })).Balance;
-        balance.ShouldBe(supposedBalance);
-    }
-
-    private async Task<Address> CreateRegiment_Use_NotAdmin()
-    {
-        var executionResult = await OracleContractStub.CreateRegiment.SendAsync(new CreateRegimentInput
-        {
-            Manager = DefaultSenderAddress,
-            IsApproveToJoin = true
-        });
-        var regimentAddress = RegimentCreated.Parser
-            .ParseFrom(executionResult.TransactionResult.Logs.First(l => l.Name == nameof(RegimentCreated))
-                .NonIndexed).RegimentAddress;
-        return regimentAddress;
-    }
 
     #endregion
 }
